refactor: 도메인 별 엔티티 관리 패키지명 domain -> entity로 변경
