+++ domain/domain-module/src/main/java/org/domainmodule/agent/entity/Agent.java
@@ -1,9 +1,14 @@
-package org.domainmodule.user.domain;
+package org.domainmodule.agent.entity;
 
+import org.domainmodule.agent.entity.type.AgentPlatform;
+import org.domainmodule.agent.entity.type.AgentType;
 import org.domainmodule.common.entity.BaseTimeEntity;
+import org.domainmodule.user.entity.User;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -17,18 +22,30 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Oauth extends BaseTimeEntity {
+public class Agent extends BaseTimeEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "oauth_id")
+	@Column(name = "agent_id")
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id")
 	private User user;
 
-	private String provider;
+	@Enumerated(EnumType.STRING)
+	private AgentPlatform platform;
 
-	private String providerId;
+	private String accessToken;
+
+	private String refreshToken;
+
+	private String accountId;
+
+	private String bio;
+
+	private Boolean autoMode;
+
+	@Enumerated(EnumType.STRING)
+	private AgentType agentType;
 }
