+++ domain/domain-module/src/main/java/org/domainmodule/agent/entity/type/AgentType.java
@@ -1,4 +1,4 @@
-package org.domainmodule.agent.domain.type;
+package org.domainmodule.agent.entity.type;
 
 public enum AgentType {
 	PERSONAL, BUSINESS
