+++ domain/domain-module/src/main/java/org/domainmodule/agent/repository/AgentRepository.java
@@ -1,6 +1,6 @@
 package org.domainmodule.agent.repository;
 
-import org.domainmodule.agent.domain.Agent;
+import org.domainmodule.agent.entity.Agent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AgentRepository extends JpaRepository<Agent, Long> {
