+++ domain/domain-module/src/main/java/org/domainmodule/post/entity/PromptHistory.java
@@ -1,15 +1,18 @@
-package org.domainmodule.user.domain;
+package org.domainmodule.post.entity;
 
 import org.domainmodule.common.entity.BaseTimeEntity;
+import org.domainmodule.post.entity.type.PostPromptType;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,16 +20,21 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RefreshToken extends BaseTimeEntity {
+public class PromptHistory extends BaseTimeEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "token_id")
+	@Column(name = "prompt_history_id")
 	private Long id;
 
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "post_id")
+	private Post post;
 
-	private String refreshToken;
+	private String prompt;
+
+	private String response;
+
+	@Enumerated(EnumType.STRING)
+	private PostPromptType promptType;
 }
