+++ domain/domain-module/src/main/java/org/domainmodule/post/entity/type/PostPromptType.java
@@ -1,4 +1,4 @@
-package org.domainmodule.post.domain.type;
+package org.domainmodule.post.entity.type;
 
 public enum PostPromptType {
 	EACH, ALL
