+++ domain/domain-module/src/main/java/org/domainmodule/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package org.domainmodule.post.repository;
 
-import org.domainmodule.post.domain.Post;
+import org.domainmodule.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
