+++ domain/domain-module/src/main/java/org/domainmodule/postgroup/entity/type/PostGroupPurpose.java
@@ -1,4 +1,4 @@
-package org.domainmodule.postgroup.domain.type;
+package org.domainmodule.postgroup.entity.type;
 
 public enum PostGroupPurpose {
 	INFORMATION, OPINION, HUMOR, MARKETING
