+++ domain/domain-module/src/main/java/org/domainmodule/postgroup/entity/type/PostGroupReference.java
@@ -1,4 +1,4 @@
-package org.domainmodule.postgroup.domain.type;
+package org.domainmodule.postgroup.entity.type;
 
 public enum PostGroupReference {
 	NONE, NEWS, IMAGE
