+++ domain/domain-module/src/main/java/org/domainmodule/postgroup/repository/PostGroupRepository.java
@@ -1,6 +1,6 @@
 package org.domainmodule.postgroup.repository;
 
-import org.domainmodule.postgroup.domain.PostGroup;
+import org.domainmodule.postgroup.entity.PostGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostGroupRepository extends JpaRepository<PostGroup, Long> {
