+++ domain/domain-module/src/main/java/org/domainmodule/rssfeed/repository/RssFeedRepository.java
@@ -1,6 +1,6 @@
 package org.domainmodule.rssfeed.repository;
 
-import org.domainmodule.rssfeed.domain.RssFeed;
+import org.domainmodule.rssfeed.entity.RssFeed;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RssFeedRepository extends JpaRepository<RssFeed, Long> {
