+++ domain/domain-module/src/main/java/org/domainmodule/user/entity/RefreshToken.java
@@ -1,18 +1,15 @@
-package org.domainmodule.post.domain;
+package org.domainmodule.user.entity;
 
 import org.domainmodule.common.entity.BaseTimeEntity;
-import org.domainmodule.post.domain.type.PostPromptType;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,21 +17,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PromptHistory extends BaseTimeEntity {
+public class RefreshToken extends BaseTimeEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "prompt_history_id")
+	@Column(name = "token_id")
 	private Long id;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "post_id")
-	private Post post;
+	@OneToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private User user;
 
-	private String prompt;
-
-	private String response;
-
-	@Enumerated(EnumType.STRING)
-	private PostPromptType promptType;
+	private String refreshToken;
 }
