+++ domain/domain-module/src/main/java/org/domainmodule/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.domainmodule.user.repository;
 
-import org.domainmodule.user.domain.User;
+import org.domainmodule.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
